pyproject.toml
[tool.poetry]
name = "brut"
version = "0.1.0"
description = ""
authors = ["Stephen Bunn <stephen@bunn.io>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{ include = "brut", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10,<3.11"
pydapper = "^0.5.3"
caribou = "^0.3.0"
pypika = "^0.48.9"
pyyaml = "^6.0"
jsonschema = "^4.17.3"
dramatiq = {extras = ["redis", "watch"], version = "^1.13.0"}
redis = "^4.4.2"
environ-config = "^23.1.0"
apscheduler = "^3.9.1.post1"
megu = {git = "https://github.com/stephen-bunn/megu.git", rev = "v2"}
url-normalize = "^1.4.3"
loguru = "^0.6.0"
praw = "^7.6.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
pre-commit = "^3.0.1"
mypy = "^0.991"
rope = "^1.7.0"

[tool.black]
line-length = 100
target-version = ["py37"]
include = '\.pyi?$'
exclude = '''
(
  /(
    \.eggs
    | \.git
    | \.mypy_cache
    | _build
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
src_paths = ["src", "test"]
multi_line_output = 3
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
